// src/lib/apis/radarr/radarr.generated.d.ts
export interface paths {
// ...
  "/api/v3/diskspace": {
    get: {
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": (components["schemas"]["DiskSpaceResource"])[];
            "application/json": (components["schemas"]["DiskSpaceResource"])[];
            "text/json": (components["schemas"]["DiskSpaceResource"])[];
          };
        };
      };
    };
  };
// ...
  "/api/v3/movie": {
    get: {
      parameters: {
        query?: {
          tmdbId?: number;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": (components["schemas"]["MovieResource"])[];
            "application/json": (components["schemas"]["MovieResource"])[];
            "text/json": (components["schemas"]["MovieResource"])[];
          };
        };
      };
    };
    post: {
      requestBody?: {
        content: {
          "application/json": components["schemas"]["MovieResource"];
          "text/json": components["schemas"]["MovieResource"];
          "application/*+json": components["schemas"]["MovieResource"];
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": components["schemas"]["MovieResource"];
            "application/json": components["schemas"]["MovieResource"];
            "text/json": components["schemas"]["MovieResource"];
          };
        };
      };
    };
  };
// ...
  "/api/v3/moviefile/{id}": {
    get: {
      parameters: {
        path: {
          id: number;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": components["schemas"]["MovieFileResource"];
            "application/json": components["schemas"]["MovieFileResource"];
            "text/json": components["schemas"]["MovieFileResource"];
          };
        };
      };
    };
    put: {
      parameters: {
        path: {
          id: string;
        };
      };
      requestBody?: {
        content: {
          "application/json": components["schemas"]["MovieFileResource"];
          "text/json": components["schemas"]["MovieFileResource"];
          "application/*+json": components["schemas"]["MovieFileResource"];
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": components["schemas"]["MovieFileResource"];
            "application/json": components["schemas"]["MovieFileResource"];
            "text/json": components["schemas"]["MovieFileResource"];
          };
        };
      };
    };
    delete: {
      parameters: {
        path: {
          id: number;
        };
      };
      responses: {
        /** @description Success */
        200: never;
      };
    };
  };
// ...
  "/api/v3/movie/lookup/tmdb": {
    get: {
      parameters: {
        query?: {
          tmdbId?: number;
        };
      };
      responses: {
        /** @description Success */
        200: never;
      };
    };
  };
// ...
  "/api/v3/queue/{id}": {
    get: {
      parameters: {
        path: {
          id: number;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": components["schemas"]["QueueResource"];
            "application/json": components["schemas"]["QueueResource"];
            "text/json": components["schemas"]["QueueResource"];
          };
        };
      };
    };
    delete: {
      parameters: {
        query?: {
          removeFromClient?: boolean;
          blocklist?: boolean;
        };
        path: {
          id: number;
        };
      };
      responses: {
        /** @description Success */
        200: never;
      };
    };
  };
// ...
  "/api/v3/queue": {
    get: {
      parameters: {
        query?: {
          includeUnknownMovieItems?: boolean;
          includeMovie?: boolean;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": components["schemas"]["QueueResourcePagingResource"];
            "application/json": components["schemas"]["QueueResourcePagingResource"];
            "text/json": components["schemas"]["QueueResourcePagingResource"];
          };
        };
      };
    };
  };
// ...
  "/api/v3/release": {
    get: {
      parameters: {
        query?: {
          movieId?: number;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "text/plain": (components["schemas"]["ReleaseResource"])[];
            "application/json": (components["schemas"]["ReleaseResource"])[];
            "text/json": (components["schemas"]["ReleaseResource"])[];
          };
        };
      };
    };
    post: {
      requestBody?: {
        content: {
          "application/json": components["schemas"]["ReleaseResource"];
          "text/json": components["schemas"]["ReleaseResource"];
          "application/*+json": components["schemas"]["ReleaseResource"];
        };
      };
      responses: {
        /** @description Success */
        200: never;
      };
    };
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
// ...
    DiskSpaceResource: {
      /** Format: int32 */
      id?: number;
      path?: string | null;
      label?: string | null;
      /** Format: int64 */
      freeSpace?: number;
      /** Format: int64 */
      totalSpace?: number;
    };
// ...
    MovieFileResource: {
      /** Format: int32 */
      id?: number;
      /** Format: int32 */
      movieId?: number;
      relativePath?: string | null;
      path?: string | null;
      /** Format: int64 */
      size?: number;
      /** Format: date-time */
      dateAdded?: string;
      sceneName?: string | null;
      /** Format: int32 */
      indexerFlags?: number;
      quality?: components["schemas"]["QualityModel"];
      customFormats?: (components["schemas"]["CustomFormatResource"])[] | null;
      mediaInfo?: components["schemas"]["MediaInfoResource"];
      originalFilePath?: string | null;
      qualityCutoffNotMet?: boolean;
      languages?: (components["schemas"]["Language"])[] | null;
      releaseGroup?: string | null;
      edition?: string | null;
    };
// ...
    MovieResource: {
      /** Format: int32 */
      id?: number;
      title?: string | null;
      originalTitle?: string | null;
      originalLanguage?: components["schemas"]["Language"];
      alternateTitles?: (components["schemas"]["AlternativeTitleResource"])[] | null;
      /** Format: int32 */
      secondaryYear?: number | null;
      /** Format: int32 */
      secondaryYearSourceId?: number;
      sortTitle?: string | null;
      /** Format: int64 */
      sizeOnDisk?: number | null;
      status?: components["schemas"]["MovieStatusType"];
      overview?: string | null;
      /** Format: date-time */
      inCinemas?: string | null;
      /** Format: date-time */
      physicalRelease?: string | null;
      /** Format: date-time */
      digitalRelease?: string | null;
      physicalReleaseNote?: string | null;
      images?: (components["schemas"]["MediaCover"])[] | null;
      website?: string | null;
      remotePoster?: string | null;
      /** Format: int32 */
      year?: number;
      hasFile?: boolean;
      youTubeTrailerId?: string | null;
      studio?: string | null;
      path?: string | null;
      /** Format: int32 */
      qualityProfileId?: number;
      monitored?: boolean;
      minimumAvailability?: components["schemas"]["MovieStatusType"];
      isAvailable?: boolean;
      folderName?: string | null;
      /** Format: int32 */
      runtime?: number;
      cleanTitle?: string | null;
      imdbId?: string | null;
      /** Format: int32 */
      tmdbId?: number;
      titleSlug?: string | null;
      rootFolderPath?: string | null;
      folder?: string | null;
      certification?: string | null;
      genres?: (string)[] | null;
      tags?: (number)[] | null;
      /** Format: date-time */
      added?: string;
      addOptions?: components["schemas"]["AddMovieOptions"];
      ratings?: components["schemas"]["Ratings"];
      movieFile?: components["schemas"]["MovieFileResource"];
      collection?: components["schemas"]["MovieCollection"];
      /** Format: float */
      popularity?: number;
    };
// ...
    QueueResource: {
      /** Format: int32 */
      id?: number;
      /** Format: int32 */
      movieId?: number | null;
      movie?: components["schemas"]["MovieResource"];
      languages?: (components["schemas"]["Language"])[] | null;
      quality?: components["schemas"]["QualityModel"];
      customFormats?: (components["schemas"]["CustomFormatResource"])[] | null;
      /** Format: double */
      size?: number;
      title?: string | null;
      /** Format: double */
      sizeleft?: number;
      timeleft?: components["schemas"]["TimeSpan"];
      /** Format: date-time */
      estimatedCompletionTime?: string | null;
      status?: string | null;
      trackedDownloadStatus?: components["schemas"]["TrackedDownloadStatus"];
      trackedDownloadState?: components["schemas"]["TrackedDownloadState"];
      statusMessages?: (components["schemas"]["TrackedDownloadStatusMessage"])[] | null;
      errorMessage?: string | null;
      downloadId?: string | null;
      protocol?: components["schemas"]["DownloadProtocol"];
      downloadClient?: string | null;
      indexer?: string | null;
      outputPath?: string | null;
    };
// ...
    ReleaseResource: {
      /** Format: int32 */
      id?: number;
      guid?: string | null;
      quality?: components["schemas"]["QualityModel"];
      customFormats?: (components["schemas"]["CustomFormatResource"])[] | null;
      /** Format: int32 */
      customFormatScore?: number;
      /** Format: int32 */
      qualityWeight?: number;
      /** Format: int32 */
      age?: number;
      /** Format: double */
      ageHours?: number;
      /** Format: double */
      ageMinutes?: number;
      /** Format: int64 */
      size?: number;
      /** Format: int32 */
      indexerId?: number;
      indexer?: string | null;
      releaseGroup?: string | null;
      subGroup?: string | null;
      releaseHash?: string | null;
      title?: string | null;
      sceneSource?: boolean;
      movieTitles?: (string)[] | null;
      languages?: (components["schemas"]["Language"])[] | null;
      approved?: boolean;
      temporarilyRejected?: boolean;
      rejected?: boolean;
      /** Format: int32 */
      tmdbId?: number;
      /** Format: int32 */
      imdbId?: number;
      rejections?: (string)[] | null;
      /** Format: date-time */
      publishDate?: string;
      commentUrl?: string | null;
      downloadUrl?: string | null;
      infoUrl?: string | null;
      downloadAllowed?: boolean;
      /** Format: int32 */
      releaseWeight?: number;
      indexerFlags?: (string)[] | null;
      edition?: string | null;
      magnetUrl?: string | null;
      infoHash?: string | null;
      /** Format: int32 */
      seeders?: number | null;
      /** Format: int32 */
      leechers?: number | null;
      protocol?: components["schemas"]["DownloadProtocol"];
      /** Format: int32 */
      movieId?: number | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}